HEEDLS-572 Split Basic accessibility tests into non-auth and auth
