+++ DigitalLearningSolutions.Web.AutomatedUiTests/AccessibilityTests/BasicAuthenticatedAccessibilityTests.cs
@@ -1,0 +1,54 @@
+namespace DigitalLearningSolutions.Web.AutomatedUiTests.AccessibilityTests
+{
+    using DigitalLearningSolutions.Web.AutomatedUiTests.TestHelpers;
+    using Xunit;
+
+    public class BasicAuthenticatedAccessibilityTests : AccessibilityTestsBase, IClassFixture<AuthenticatedAccessibilityTestsFixture<Startup>>
+    {
+        public BasicAuthenticatedAccessibilityTests(AuthenticatedAccessibilityTestsFixture<Startup> fixture) : base(fixture) { }
+        
+        [Theory]
+        [InlineData("/MyAccount", "My account")]
+        [InlineData("/MyAccount/EditDetails", "Edit details")]
+        [InlineData("/TrackingSystem/Centre/Administrators", "Centre administrators")]
+        [InlineData("/TrackingSystem/Centre/Dashboard", "Centre dashboard")]
+        [InlineData("/TrackingSystem/Centre/Ranking", "Centre ranking")]
+        [InlineData("/TrackingSystem/Centre/ContractDetails", "Contract details")]
+        [InlineData("/TrackingSystem/Centre/Configuration", "Centre configuration")]
+        [InlineData("/TrackingSystem/Centre/Configuration/EditCentreDetails", "Edit centre details")]
+        [InlineData("/TrackingSystem/Centre/Configuration/EditCentreManagerDetails", "Edit centre manager details")]
+        [InlineData(
+            "/TrackingSystem/Centre/Configuration/EditCentreWebsiteDetails",
+            "Edit centre content on DLS website"
+        )]
+        [InlineData("/TrackingSystem/Centre/Configuration/RegistrationPrompts", "Manage delegate registration prompts")]
+        [InlineData(
+            "/TrackingSystem/Centre/Configuration/RegistrationPrompts/1/Remove",
+            "Remove delegate registration prompt"
+        )]
+        [InlineData("/TrackingSystem/Centre/Reports", "Centre reports")]
+        [InlineData("/TrackingSystem/Centre/TopCourses", "Top courses")]
+        [InlineData("/TrackingSystem/CourseSetup", "Centre course setup")]
+        [InlineData("/TrackingSystem/CourseSetup/10716/AdminFields", "Manage course admin fields")]
+        [InlineData("/TrackingSystem/CourseSetup/10716/Content", "Course content")]
+        [InlineData("/TrackingSystem/CourseSetup/10716/Manage", "Level 1 - Microsoft Excel 2010 - Inductions")]
+        [InlineData("/TrackingSystem/Delegates/All", "Delegates")]
+        [InlineData("/TrackingSystem/Delegates/Groups", "Groups")]
+        [InlineData("/TrackingSystem/Delegates/View/3", "xxxx xxxxxx")]
+        [InlineData("/TrackingSystem/Delegates/Approve", "Approve delegate registrations")]
+        [InlineData("/TrackingSystem/Delegates/BulkUpload", "Bulk upload/update delegates")]
+        [InlineData("/NotificationPreferences", "Notification preferences")]
+        [InlineData("/NotificationPreferences/Edit/AdminUser", "Update notification preferences")]
+        [InlineData("/NotificationPreferences/Edit/DelegateUser", "Update notification preferences")]
+        [InlineData("/ChangePassword", "Change password")]
+        [InlineData("/TrackingSystem/Support", "Support")]
+        public void Authenticated_page_has_no_accessibility_errors(string url, string pageTitle)
+        {
+            // when
+            Driver.Navigate().GoToUrl(BaseUrl + url);
+
+            // then
+            AnalyzePageHeadingAndAccessibility(pageTitle);
+        }
+    }
+}
